Protest Classifer For KUSRP-2020 Task 1
This mini-project consist of 2 main parts.

1- Extraction

In the first part, news from The Hindu are extracted from their given
URL and saved in a csv.

You can check extract_texts.ipynb notebook to see how they are extracted
and extracted_data.csv to see raw data.

2- EDA_Preporcessing_Models

On the second part, first of all, raw data is being cleaned and couple points are
analyzed on the cleaned data.

Then, two diffrent approach and two different algorithm for each
approach are applied and results are shown as "Test Scores".